flex fix for .disqus and .wrapper-footer
In Safari (and in Chrome to a lesser degree), setting flex-direction on
a parent container and not specifically setting flex-basis and
flex-grow can cause layout bugs. I believe this is a result of a poor
implementation of flex-direction by Safari/WebKit. But the fix works
cross-browser.

Tested in Safari 9.0.2 on OS X, Safari 9.0.2 on iOS, latest Firefox,
latest Chrome.